% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_42CBTC}
\alias{GeneralMarketData_42CBTC}
\title{42coin to Bitcoin}
\usage{
GeneralMarketData_42CBTC()
}
\arguments{
\item{42COIN}{to BTC - Realtime}
}
\description{
This function allows you to get general market data on 42coin to BITCOIN
}
\examples{
GeneralMarketData_42CBTC()
}
\keyword{42coin}

