% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_ADTLTC}
\alias{GeneralMarketData_ADTLTC}
\title{AndroidsTokensV2  to Litecoin}
\usage{
GeneralMarketData_ADTLTC()
}
\description{
This function allows you to get general market data on AndroidsTokensV2 to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_ADTLTC()
}
}
\keyword{AndroidsTokensV2}

