% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_ANCLTC}
\alias{GeneralMarketData_ANCLTC}
\title{AnonCoin  to Litecoin}
\usage{
GeneralMarketData_ANCLTC()
}
\description{
This function allows you to get general market data on AnonCoin to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_ANCLTC()
}
}
\keyword{AnonCoin}

