% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_ASCLTC}
\alias{GeneralMarketData_ASCLTC}
\title{AsicCoin   to Litecoin}
\usage{
GeneralMarketData_ASCLTC()
}
\arguments{
\item{AsicCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on AsicCoin to LITECOIN
}
\examples{
GeneralMarketData_ASCLTC()
}
\keyword{AsicCoin}

