% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_AURLTC}
\alias{GeneralMarketData_AURLTC}
\title{AuroraCoin to Litecoin}
\usage{
GeneralMarketData_AURLTC()
}
\arguments{
\item{AuroraCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on AuroraCoin to LITECOIN
}
\examples{
GeneralMarketData_AURLTC()
}
\keyword{AuroraCoin}

