% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_BATLTC}
\alias{GeneralMarketData_BATLTC}
\title{BatCoin to Litecoin}
\usage{
GeneralMarketData_BATLTC()
}
\arguments{
\item{BatCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on BatCoin to LITECOIN
}
\examples{
GeneralMarketData_BATLTC()
}
\keyword{BatCoin}

