% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_BCLTC}
\alias{GeneralMarketData_BCLTC}
\title{BlackCoin to Litecoin}
\usage{
GeneralMarketData_BCLTC()
}
\description{
This function allows you to get general market data on BlackCoin  to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_BATLTC()
}
}
\keyword{BlackCoin}

