% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_BTCUSD}
\alias{GeneralMarketData_BTCUSD}
\title{BITCOIN to USD}
\usage{
GeneralMarketData_BTCUSD()
}
\description{
This function allows you to get general market data on Bitcoin to US Dollars
}
\examples{
\dontrun{
GeneralMarketData_BTCUSD()
}
}
\keyword{bitcoin}

