% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_CAIxLTC}
\alias{GeneralMarketData_CAIxLTC}
\title{CAIx to Litecoin}
\usage{
GeneralMarketData_CAIxLTC()
}
\arguments{
\item{CAIx}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on CAIx to LITECOIN
}
\examples{
GeneralMarketData_CAIxLTC()
}
\keyword{CAIx}

