% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_CGBLTC}
\alias{GeneralMarketData_CGBLTC}
\title{CryptogenicBullion to Litecoin}
\usage{
GeneralMarketData_CGBLTC()
}
\arguments{
\item{CryptogenicBullion}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on CryptogenicBullion to LITECOIN
}
\examples{
GeneralMarketData_CGBLTC()
}
\keyword{CryptogenicBullion}

