% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DOGEBTC}
\alias{GeneralMarketData_DOGEBTC}
\title{Dogecoin to Bitcoin}
\usage{
GeneralMarketData_DOGEBTC()
}
\arguments{
\item{DOGECOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general market data on Dogecoin to BITCOIN
}
\examples{
GeneralMarketData_DOGEBTC()
}
\keyword{Dogecoin}

