% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DOGELTC}
\alias{GeneralMarketData_DOGELTC}
\title{Dogecoin to Litecoin}
\usage{
GeneralMarketData_DOGELTC()
}
\description{
This function allows you to get general market data on Dogecoin to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_DOGELTC()
}
}
\keyword{Dogecoin}

