% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_DOGEUSD}
\alias{GeneralMarketData_DOGEUSD}
\title{DOGECOIN to USD}
\usage{
GeneralMarketData_DOGEUSD()
}
\arguments{
\item{DOGECOIN}{to USD - Realtime}
}
\description{
This function allows you get general market data on Dogecoin to US Dollars
}
\examples{
GeneralMarketData_DOGEUSD()
}
\keyword{doge}

