% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DRKBTC}
\alias{GeneralMarketData_DRKBTC}
\title{Darkcoin to Bitcoin}
\usage{
GeneralMarketData_DRKBTC()
}
\description{
This function allows you to get general market data on Darkcoin to BITCOIN
}
\examples{
\dontrun{
GeneralMarketData_DRKBTC()
}
}
\keyword{Darkcoin}

