% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DRKUSD}
\alias{GeneralMarketData_DRKUSD}
\title{Darkcoin to USD}
\usage{
GeneralMarketData_DRKUSD()
}
\description{
This function allows you to get general market data on Darkcoin to US Dollars
}
\examples{
\dontrun{
GeneralMarketData_DRKUSD()
}
}
\keyword{Darkcoin}

