% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_FTCBTC}
\alias{GeneralMarketData_FTCBTC}
\title{Feathercoin to Bitcoin}
\usage{
GeneralMarketData_FTCBTC()
}
\arguments{
\item{FEATHERCOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general market data on Feathercoin to BITCOIN
}
\examples{
GeneralMarketData_FTCBTC()
}
\keyword{Feathercoin}

