% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_FTCUSD}
\alias{GeneralMarketData_FTCUSD}
\title{Feathercoin to USD}
\usage{
GeneralMarketData_FTCUSD()
}
\arguments{
\item{FEATHERCOIN}{to USD - Realtime}
}
\description{
This function allows you to get general market data on Feathercoin to US Dollars
}
\examples{
GeneralMarketData_FTCUSD()
}
\keyword{feathercoin}

