% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_LTCBTC}
\alias{GeneralMarketData_LTCBTC}
\title{Litecoin to Bitcoin}
\usage{
GeneralMarketData_LTCBTC()
}
\description{
This function allows you to get general market data on Litecoin to BITCOIN
}
\examples{
\dontrun{
GeneralMarketData_LTCBTC()
}
}
\keyword{Litecoin}

