% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_NXTBTC}
\alias{GeneralMarketData_NXTBTC}
\title{NXTcoin to Bitcoin}
\usage{
GeneralMarketData_NXTBTC()
}
\arguments{
\item{NXTCOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general market data on NXTcoin to BITCOIN
}
\examples{
GeneralMarketData_NXTBTC()
}
\keyword{NXTcoin}

