% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_PCCUSD}
\alias{GeneralMarketData_PCCUSD}
\title{Peercoin to USD}
\usage{
GeneralMarketData_PCCUSD()
}
\description{
This function allows you to get general market data on Peercoin to US Dollars
}
\examples{
\dontrun{
GeneralMarketData_PCCUSD()
}
}
\keyword{Peercoin}

