% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_RDDUSD}
\alias{GeneralMarketData_RDDUSD}
\title{ReddCoin to USD}
\usage{
GeneralMarketData_RDDUSD()
}
\arguments{
\item{ReddCoin}{to USD  - Realtime}
}
\description{
This function allows you to get general market data on ReddCoin to US Dollars
}
\examples{
GeneralMarketData_RDDUSD()
}
\keyword{ReddCoin}

