% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_UNOBTC}
\alias{GeneralMarketData_UNOBTC}
\title{Unobtanium to Bitcoin}
\usage{
GeneralMarketData_UNOBTC()
}
\description{
This function allows you to get general market data on Unobtaniumto US Dollars
}
\examples{
\dontrun{
GeneralMarketData_UNOBTC()
}
}
\keyword{Unobtanium}

