% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_XRPUSD}
\alias{GeneralMarketData_XRPUSD}
\title{Ripple to USD}
\usage{
GeneralMarketData_XRPUSD()
}
\arguments{
\item{Ripple}{to USD  - Realtime}
}
\description{
This function allows you to get general market data on Ripple to US Dollars
}
\examples{
GeneralMarketData_XRPUSD()
}
\keyword{Ripple}

