% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_42CBTC}
\alias{OrderBookData_42CBTC}
\title{42coin to Bitcoin}
\usage{
OrderBookData_42CBTC()
}
\arguments{
\item{42COIN}{to BTC - Realtime}
}
\description{
This function allows you to get general orderbook data on 42coin to BITCOIN
}
\examples{
OrderBookData_42CBTC()
}
\keyword{42coin}

