% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_ADTLTC}
\alias{OrderBookData_ADTLTC}
\title{AndroidsTokensV2  to Litecoin}
\usage{
OrderBookData_ADTLTC()
}
\arguments{
\item{AndroidsTokensV2}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on AndroidsTokensV2 to LITECOIN
}
\examples{
OrderBookData_ADTLTC()
}
\keyword{AndroidsTokensV2}

