% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_ANCLTC}
\alias{OrderBookData_ANCLTC}
\title{AnonCoin  to Litecoin}
\usage{
OrderBookData_ANCLTC()
}
\arguments{
\item{AnonCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on AnonCoin to LITECOIN
}
\examples{
OrderBookData_ANCLTC()
}
\keyword{AnonCoin}

