% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_AURLTC}
\alias{OrderBookData_AURLTC}
\title{AuroraCoin to Litecoin}
\usage{
OrderBookData_AURLTC()
}
\arguments{
\item{AuroraCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on AuroraCoin to LITECOIN
}
\examples{
OrderBookData_AURLTC()
}
\keyword{AuroraCoin}

