% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_BATLTC}
\alias{OrderBookData_BATLTC}
\title{BatCoin to Litecoin}
\usage{
OrderBookData_BATLTC()
}
\arguments{
\item{BatCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on BatCoin to LITECOIN
}
\examples{
OrderBookData_BATLTC()
}
\keyword{BatCoin}

