% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_BCLTC}
\alias{OrderBookData_BCLTC}
\title{BlackCoin to Litecoin}
\usage{
OrderBookData_BCLTC()
}
\arguments{
\item{BlackCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on BlackCoin  to LITECOIN
}
\examples{
OrderBookData_BATLTC()
}
\keyword{BlackCoin}

