% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_BTCUSD}
\alias{OrderBookData_BTCUSD}
\title{BITCOIN to USD}
\usage{
OrderBookData_BTCUSD()
}
\arguments{
\item{BITCOIN}{to USD - Realtime}
}
\description{
This function allows you to get general orderbook data on Bitcoin to US Dollars
}
\examples{
OrderBookData_BTCUSD()
}
\keyword{bitcoin}

