% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_CAIxLTC}
\alias{OrderBookData_CAIxLTC}
\title{CAIx to Litecoin}
\usage{
OrderBookData_CAIxLTC()
}
\arguments{
\item{CAIx}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on CAIx to LITECOIN
}
\examples{
OrderBookData_CAIxLTC()
}
\keyword{CAIx}

