% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_CGBLTC}
\alias{OrderBookData_CGBLTC}
\title{CryptogenicBullion to Litecoin}
\usage{
OrderBookData_CGBLTC()
}
\arguments{
\item{CryptogenicBullion}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on CryptogenicBullion to LITECOIN
}
\examples{
OrderBookData_CGBLTC()
}
\keyword{CryptogenicBullion}

