% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_DOGEBTC}
\alias{OrderBookData_DOGEBTC}
\title{Dogecoin to Bitcoin}
\usage{
OrderBookData_DOGEBTC()
}
\arguments{
\item{DOGECOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general orderbook data on Dogecoin to BITCOIN
}
\examples{
OrderBookData_DOGEBTC()
}
\keyword{Dogecoin}

