% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_DOGELTC}
\alias{OrderBookData_DOGELTC}
\title{Dogecoin to Litecoin}
\usage{
OrderBookData_DOGELTC()
}
\arguments{
\item{Dogecoin}{to LTC - Realtime}
}
\description{
This function allows you to get general orderbook data on Dogecoin to LITECOIN
}
\examples{
OrderBookData_DOGELTC()
}
\keyword{Dogecoin}

