% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_DOGEUSD}
\alias{OrderBookData_DOGEUSD}
\title{DOGECOIN to USD}
\usage{
OrderBookData_DOGEUSD()
}
\arguments{
\item{DOGECOIN}{to USD - Realtime}
}
\description{
This function allows you to get general orderbook data on Dogecoin to US Dollars
}
\examples{
OrderBookData_DOGEUSD()
}
\keyword{doge}

