% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_DRKBTC}
\alias{OrderBookData_DRKBTC}
\title{Darkcoin to Bitcoin}
\usage{
OrderBookData_DRKBTC()
}
\arguments{
\item{DARKCOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general orderbook data on Darkcoin to BITCOIN
}
\examples{
OrderBookData_DRKBTC()
}
\keyword{Darkcoin}

