% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_DRKUSD}
\alias{OrderBookData_DRKUSD}
\title{Darkcoin to USD}
\usage{
OrderBookData_DRKUSD()
}
\arguments{
\item{DARKCOIN}{to USD - Realtime}
}
\description{
This function allows you to get general orderbook data on Darkcoin to US Dollars
}
\examples{
OrderBookData_DRKUSD()
}
\keyword{Darkcoin}

