% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_FTCBTC}
\alias{OrderBookData_FTCBTC}
\title{Feathercoin to Bitcoin}
\usage{
OrderBookData_FTCBTC()
}
\arguments{
\item{FEATHERCOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general orderbook data on Feathercoin to BITCOIN
}
\examples{
OrderBookData_FTCBTC()
}
\keyword{Feathercoin}

