% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_FTCUSD}
\alias{OrderBookData_FTCUSD}
\title{Feathercoin to USD}
\usage{
OrderBookData_FTCUSD()
}
\arguments{
\item{FEATHERCOIN}{to USD - Realtime}
}
\description{
This function allows you to get general orderbook data on Feathercoin to US Dollars
}
\examples{
OrderBookData_FTCUSD()
}
\keyword{feathercoin}

