% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_LTCBTC}
\alias{OrderBookData_LTCBTC}
\title{Litecoin to Bitcoin}
\usage{
OrderBookData_LTCBTC()
}
\arguments{
\item{LITECOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general orderbook data on Litecoin to BITCOIN
}
\examples{
OrderBookData_LTCBTC()
}
\keyword{Litecoin}

