% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_LTCUSD}
\alias{OrderBookData_LTCUSD}
\title{Litecoin to USD}
\usage{
OrderBookData_LTCUSD()
}
\arguments{
\item{LITECOIN}{to USD - Realtime}
}
\description{
This function allows you to get general orderbook data on Litecoin to US Dollars
}
\examples{
OrderBookData_LTCUSD()
}
\keyword{litecoin}

