% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_NXTBTC}
\alias{OrderBookData_NXTBTC}
\title{NXTcoin to Bitcoin}
\usage{
OrderBookData_NXTBTC()
}
\arguments{
\item{NXTCOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general orderbook data on NXTcoin to BITCOIN
}
\examples{
OrderBookData_NXTBTC()
}
\keyword{NXTcoin}

