% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_PCCUSD}
\alias{OrderBookData_PCCUSD}
\title{Peercoin to USD}
\usage{
OrderBookData_PCCUSD()
}
\arguments{
\item{Peercoin}{to USD  - Realtime}
}
\description{
This function allows you to get general orderbook data on Peercoin to US Dollars
}
\examples{
OrderBookData_PCCUSD()
}
\keyword{Peercoin}

