% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_RDDUSD}
\alias{OrderBookData_RDDUSD}
\title{ReddCoin to USD}
\usage{
OrderBookData_RDDUSD()
}
\arguments{
\item{ReddCoin}{to USD  - Realtime}
}
\description{
This function allows you to get general orderbook data on ReddCoin to US Dollars
}
\examples{
OrderBookData_RDDUSD()
}
\keyword{ReddCoin}

