% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_UNOBTC}
\alias{OrderBookData_UNOBTC}
\title{Unobtanium to Bitcoin}
\usage{
OrderBookData_UNOBTC()
}
\arguments{
\item{Unobtanium}{to Bitcoin  - Realtime}
}
\description{
This function allows you to get general orderbook data on Unobtaniumto US Dollars
}
\examples{
OrderBookData_UNOBTC()
}
\keyword{Unobtanium}

