% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderbookdata.R
\name{OrderBookData_XRPUSD}
\alias{OrderBookData_XRPUSD}
\title{Ripple to USD}
\usage{
OrderBookData_XRPUSD()
}
\arguments{
\item{Ripple}{to USD  - Realtime}
}
\description{
This function allows you to get general orderbook data on Ripple to US Dollars
}
\examples{
OrderBookData_XRPUSD()
}
\keyword{Ripple}

