\name{RCryptsy-package}
\alias{RCryptsy-package}
\alias{RCryptsy}
\docType{package}
\title{
Access to the Cryptsy Crypto-Currency Exchange Public Market Information API via R (RCryptsy)
}
\description{
Basic Public API for the Crypto-Currency exchange Cryptsy.
}
\details{
\tabular{ll}{
Package: \tab RCryptsy\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2015-07-28\cr
License: \tab GPL-3\cr
}
This package provides access to the Cryptsy Crypto-Currency Exchange API via R and calls for market data.
}
\author{
William Kyle Hamilton

Maintainer: William Kyle Hamilton <kyle.hamilton@gmail.com>
}

\keyword{package}