% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_42CBTC}
\alias{GeneralMarketData_42CBTC}
\title{42coin to Bitcoin}
\usage{
GeneralMarketData_42CBTC()
}
\description{
This function allows you to get general market data on 42coin to BITCOIN.
}
\examples{
GeneralMarketData_42CBTC()
}
\keyword{42coin}

