% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_42CBTC}
\alias{GeneralMarketData_42CBTC}
\title{42coin to Bitcoin}
\usage{
GeneralMarketData_42CBTC()
}
\arguments{
\item{42COIN}{to BTC - Realtime}
}
\description{
This function allows you get general market data on 42coin to BITCOIN
}
\examples{
GeneralMarketData_42CBTC()
}
\keyword{42coin}

