% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_42CBTC}
\alias{GeneralMarketData_42CBTC}
\title{42coin to Bitcoin}
\usage{
GeneralMarketData_42CBTC()
}
\description{
This function allows you to get general market data on 42coin to BITCOIN
}
\examples{
\dontrun{
GeneralMarketData_42CBTC()
}
}

