% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_ADTLTC}
\alias{GeneralMarketData_ADTLTC}
\title{AndroidsTokensV2  to Litecoin}
\usage{
GeneralMarketData_ADTLTC()
}
\arguments{
\item{AndroidsTokensV2}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on AndroidsTokensV2 to LITECOIN
}
\examples{
GeneralMarketData_ADTLTC()
}
\keyword{AndroidsTokensV2}

