% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_ANCLTC}
\alias{GeneralMarketData_ANCLTC}
\title{AnonCoin  to Litecoin}
\usage{
GeneralMarketData_ANCLTC()
}
\arguments{
\item{AnonCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on AnonCoin to LITECOIN
}
\examples{
GeneralMarketData_ANCLTC()
}
\keyword{AnonCoin}

