% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_ASCLTC}
\alias{GeneralMarketData_ASCLTC}
\title{AsicCoin to Litecoin}
\usage{
GeneralMarketData_ASCLTC()
}
\description{
This function allows you to get general market data on AsicCoin to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_ASCLTC()
}
}
\keyword{AsicCoin}

