% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_AURLTC}
\alias{GeneralMarketData_AURLTC}
\title{AuroraCoin to Litecoin}
\usage{
GeneralMarketData_AURLTC()
}
\description{
This function allows you to get general market data on AuroraCoin to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_AURLTC()
}
}
\keyword{AuroraCoin}

