% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_BATLTC}
\alias{GeneralMarketData_BATLTC}
\title{BatCoin to Litecoin}
\usage{
GeneralMarketData_BATLTC()
}
\description{
This function allows you to get general market data on BatCoin to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_BATLTC()
}
}
\keyword{BatCoin}

