% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_BCLTC}
\alias{GeneralMarketData_BCLTC}
\title{BlackCoin to Litecoin}
\usage{
GeneralMarketData_BCLTC()
}
\arguments{
\item{BlackCoin}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on BlackCoin  to LITECOIN
}
\examples{
GeneralMarketData_BATLTC()
}
\keyword{BlackCoin}

