% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_BTCUSD}
\alias{GeneralMarketData_BTCUSD}
\title{BITCOIN to USD}
\usage{
GeneralMarketData_BTCUSD()
}
\arguments{
\item{BITCOIN}{to USD - Realtime}
}
\description{
This function allows you to get general market data on Bitcoin to US Dollars
}
\examples{
GeneralMarketData_BTCUSD()
}
\keyword{bitcoin}

