% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_BTCUSD}
\alias{GeneralMarketData_BTCUSD}
\title{BITCOIN to USD}
\usage{
GeneralMarketData_BTCUSD()
}
\description{
This function allows you to get general market data on Bitcoin to US Dollars.
}
\examples{
GeneralMarketData_BTCUSD()
}
\keyword{bitcoin}

