% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_CAIxLTC}
\alias{GeneralMarketData_CAIxLTC}
\title{CAIx to Litecoin}
\usage{
GeneralMarketData_CAIxLTC()
}
\description{
This function allows you to get general market data on CAIx to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_CAIxLTC()
}
}
\keyword{CAIx}

