% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_CGBLTC}
\alias{GeneralMarketData_CGBLTC}
\title{CryptogenicBullion to Litecoin}
\usage{
GeneralMarketData_CGBLTC()
}
\description{
This function allows you to get general market data on CryptogenicBullion to LITECOIN
}
\examples{
\dontrun{
GeneralMarketData_CGBLTC()
}
}
\keyword{CryptogenicBullion}

