% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_DOGEBTC}
\alias{GeneralMarketData_DOGEBTC}
\title{Dogecoin to Bitcoin}
\usage{
GeneralMarketData_DOGEBTC()
}
\description{
This function allows you get general market data on Dogecoin to BITCOIN
}
\examples{
GeneralMarketData_DOGEBTC()
}
\keyword{Dogecoin}

