% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DOGEBTC}
\alias{GeneralMarketData_DOGEBTC}
\title{Dogecoin to Bitcoin}
\usage{
GeneralMarketData_DOGEBTC()
}
\description{
This function allows you to get general market data on Dogecoin to BITCOIN
}
\examples{
\dontrun{
GeneralMarketData_DOGEBTC()
}
}
\keyword{Dogecoin}

