% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_DOGELTC}
\alias{GeneralMarketData_DOGELTC}
\title{Dogecoin to Litecoin}
\usage{
GeneralMarketData_DOGELTC()
}
\arguments{
\item{Dogecoin}{to LTC - Realtime}
}
\description{
This function allows you get general market data on Dogecoin to LITECOIN
}
\examples{
GeneralMarketData_DOGELTC()
}
\keyword{Dogecoin}

