% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DOGELTC}
\alias{GeneralMarketData_DOGELTC}
\title{Dogecoin to Litecoin}
\usage{
GeneralMarketData_DOGELTC()
}
\arguments{
\item{Dogecoin}{to LTC - Realtime}
}
\description{
This function allows you to get general market data on Dogecoin to LITECOIN
}
\examples{
GeneralMarketData_DOGELTC()
}
\keyword{Dogecoin}

