% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_DOGEUSD}
\alias{GeneralMarketData_DOGEUSD}
\title{DOGECOIN to USD}
\usage{
GeneralMarketData_DOGEUSD()
}
\description{
This function allows you to get general market data on Dogecoin to US Dollars.
}
\examples{
GeneralMarketData_DOGEUSD()
}
\keyword{doge}

