% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_DRKBTC}
\alias{GeneralMarketData_DRKBTC}
\title{Darkcoin to Bitcoin}
\usage{
GeneralMarketData_DRKBTC()
}
\arguments{
\item{DARKCOIN}{to BTC - Realtime}
}
\description{
This function allows you get general market data on Darkcoin to BITCOIN
}
\examples{
GeneralMarketData_DRKBTC()
}
\keyword{Darkcoin}

