% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DRKBTC}
\alias{GeneralMarketData_DRKBTC}
\title{Darkcoin to Bitcoin}
\usage{
GeneralMarketData_DRKBTC()
}
\arguments{
\item{DARKCOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general market data on Darkcoin to BITCOIN
}
\examples{
GeneralMarketData_DRKBTC()
}
\keyword{Darkcoin}

