% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_DRKUSD}
\alias{GeneralMarketData_DRKUSD}
\title{Darkcoin to USD}
\usage{
GeneralMarketData_DRKUSD()
}
\arguments{
\item{DARKCOIN}{to USD - Realtime}
}
\description{
This function allows you to get general market data on Darkcoin to US Dollars
}
\examples{
GeneralMarketData_DRKUSD()
}
\keyword{Darkcoin}

