% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_DRKUSD}
\alias{GeneralMarketData_DRKUSD}
\title{Darkcoin to USD}
\usage{
GeneralMarketData_DRKUSD()
}
\description{
This function allows you to get general market data on Darkcoin to US Dollars.
}
\examples{
GeneralMarketData_DRKUSD()
}
\keyword{Darkcoin}

