% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_FTCBTC}
\alias{GeneralMarketData_FTCBTC}
\title{Feathercoin to Bitcoin}
\usage{
GeneralMarketData_FTCBTC()
}
\description{
This function allows you to get general market data on Feathercoin to BITCOIN
}
\examples{
\dontrun{
GeneralMarketData_FTCBTC()
}
}
\keyword{Feathercoin}

