% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_FTCBTC}
\alias{GeneralMarketData_FTCBTC}
\title{Feathercoin to Bitcoin}
\usage{
GeneralMarketData_FTCBTC()
}
\description{
This function allows you get general market data on Feathercoin to BITCOIN
}
\examples{
GeneralMarketData_FTCBTC()
}
\keyword{Feathercoin}

