% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_FTCUSD}
\alias{GeneralMarketData_FTCUSD}
\title{Feathercoin to USD}
\usage{
GeneralMarketData_FTCUSD()
}
\description{
This function allows you to get general market data on Feathercoin to US Dollars
}
\examples{
\dontrun{
GeneralMarketData_FTCUSD()
}
}
\keyword{feathercoin}

