% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_FTCUSD}
\alias{GeneralMarketData_FTCUSD}
\title{Feathercoin to USD}
\usage{
GeneralMarketData_FTCUSD()
}
\description{
This function allows you get general market data on Feathercoin to US Dollars
}
\examples{
GeneralMarketData_FTCUSD()
}
\keyword{feathercoin}

