% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_LTCBTC}
\alias{GeneralMarketData_LTCBTC}
\title{Litecoin to Bitcoin}
\usage{
GeneralMarketData_LTCBTC()
}
\arguments{
\item{LITECOIN}{to BTC - Realtime}
}
\description{
This function allows you to get general market data on Litecoin to BITCOIN
}
\examples{
GeneralMarketData_LTCBTC()
}
\keyword{Litecoin}

