% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_LTCBTC}
\alias{GeneralMarketData_LTCBTC}
\title{Litecoin to Bitcoin}
\usage{
GeneralMarketData_LTCBTC()
}
\description{
This function allows you get general market data on Litecoin to BITCOIN
}
\examples{
GeneralMarketData_LTCBTC()
}
\keyword{Litecoin}

