% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_LTCUSD}
\alias{GeneralMarketData_LTCUSD}
\title{Litecoin to USD}
\usage{
GeneralMarketData_LTCUSD()
}
\description{
This function allows you to get general market data on Litecoin to US Dollars.
}
\examples{
GeneralMarketData_LTCUSD()
}
\keyword{litecoin}

