% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_LTCUSD}
\alias{GeneralMarketData_LTCUSD}
\title{Litecoin to USD}
\usage{
GeneralMarketData_LTCUSD()
}
\arguments{
\item{LITECOIN}{to USD - Realtime}
}
\description{
This function allows you get general market data on Litecoin to US Dollars
}
\examples{
GeneralMarketData_LTCUSD()
}
\keyword{litecoin}

