% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralMarketData_NXTBTC}
\alias{GeneralMarketData_NXTBTC}
\title{NXTcoin to Bitcoin}
\usage{
GeneralMarketData_NXTBTC()
}
\arguments{
\item{NXTCOIN}{to BTC - Realtime}
}
\description{
This function allows you get general market data on NXTcoin to BITCOIN
}
\examples{
GeneralMarketData_NXTBTC()
}
\keyword{NXTcoin}

