% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_NXTBTC}
\alias{GeneralMarketData_NXTBTC}
\title{NXTcoin to Bitcoin}
\usage{
GeneralMarketData_NXTBTC()
}
\description{
This function allows you to get general market data on NXTcoin to BITCOIN
}
\examples{
\dontrun{
GeneralMarketData_NXTBTC()
}
}
\keyword{NXTcoin}

