% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_PCCUSD}
\alias{GeneralMarketData_PCCUSD}
\title{Peercoin to USD}
\usage{
GeneralMarketData_PCCUSD()
}
\arguments{
\item{Peercoin}{to USD  - Realtime}
}
\description{
This function allows you to get general market data on Peercoin to US Dollars
}
\examples{
GeneralMarketData_PCCUSD()
}
\keyword{Peercoin}

