% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_UNOBTC}
\alias{GeneralMarketData_UNOBTC}
\title{Unobtanium to Bitcoin}
\usage{
GeneralMarketData_UNOBTC()
}
\arguments{
\item{Unobtanium}{to Bitcoin  - Realtime}
}
\description{
This function allows you to get general market data on Unobtaniumto US Dollars
}
\examples{
GeneralMarketData_UNOBTC()
}
\keyword{Unobtanium}

