% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCryptsy.R
\name{GeneralMarketData_XRPUSD}
\alias{GeneralMarketData_XRPUSD}
\title{Ripple to USD}
\usage{
GeneralMarketData_XRPUSD()
}
\description{
This function allows you to get general market data on Ripple to US Dollars
}
\examples{
\dontrun{
GeneralMarketData_XRPUSD()
}
}
\keyword{Ripple}

