\name{RCryptsy-package}
\alias{RCryptsy-package}
\alias{RCryptsy}
\docType{package}
\title{
Access to Cryptsy Crypto-Currency Exchange Public Infomation API via R (RCryptsy)
}
\description{
Basic Public API for the Crypto-Currency exchange Cryptsy.
}
\details{
\tabular{ll}{
Package: \tab RCryptsy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-05\cr
License: \tab GPL-3\cr
}
This package provides access to the Cryptsy Crypto-Currency Exchange API via R and calls for market data.
}
\author{
William Kyle Hamilton

Maintainer: William Kyle Hamilton <kyle.hamilton@gmail.com>
}

\keyword{ package }

\examples{
GeneralMarketDataAll() will gather all the data from all of the markets
GeneralMarketData_BTCUSD will gather all the data from the Bitcoin US Dollar market
}
